Separate out gossip_message_validator module
This module will later be fulfilled with a base gossip message validator
`GossipMessageValidator` which is shared in both system and core domain.
